serving/requirements.txt
annotated-types==0.7.0
blinker @ file:///C:/b/abs_b1i87khtob/croot/blinker_1737448732095/work
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click @ file:///C:/b/abs_42rbjngfys/croot/click_1744271606983/work
colorama @ file:///C:/b/abs_a9ozq0l032/croot/colorama_1672387194846/work
filelock @ file:///C:/b/abs_758187j281/croot/filelock_1744281404850/work
Flask @ file:///C:/b/abs_a6m37uz578/croot/flask_1737454335662/work
gmpy2 @ file:///C:/b/abs_d8ki0o0h97/croot/gmpy2_1738085498525/work
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
httplib2==0.22.0
idna==3.10
importlib_metadata @ file:///C:/b/abs_9d0lp1ysmt/croot/importlib_metadata-suite_1732633510972/work
itsdangerous @ file:///C:/b/abs_c4vwgdr5yn/croot/itsdangerous_1716533399914/work
Jinja2 @ file:///C:/b/abs_920kup4e6u/croot/jinja2_1741711580669/work
MarkupSafe @ file:///C:/b/abs_a0ma7ge0jc/croot/markupsafe_1738584052792/work
mkl-service==2.4.0
mkl_fft @ file:///C:/Users/dev-admin/mkl/mkl_fft_1730823082242/work
mkl_random @ file:///C:/Users/dev-admin/mkl/mkl_random_1730822522280/work
mpmath @ file:///C:/b/abs_7833jrbiox/croot/mpmath_1690848321154/work
networkx @ file:///C:/b/abs_3bxnu56g9d/croot/networkx_1717597507456/work
numpy @ file:///C:/b/abs_c1ywpu18ar/croot/numpy_and_numpy_base_1708638681471/work/dist/numpy-1.26.4-cp39-cp39-win_amd64.whl#sha256=f284633067fde4bb078a09085a1c9e99a3218c129f5312b7b3d76c9340442a25
pillow @ file:///C:/b/abs_b06nvnir5m/croot/pillow_1752524543576/work
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dotenv @ file:///C:/b/abs_71cpoh9hpg/croot/python-dotenv_1745613639902/work
PyYAML @ file:///C:/b/abs_14xkfs39bx/croot/pyyaml_1728657968772/work
requests==2.32.5
rsa==4.9.1
sympy @ file:///C:/b/abs_b4u17p23yg/croot/sympy_1738108511395/work
torch==2.5.1
torchaudio==2.5.1
torchvision @ file:///C:/b/abs_d6bzbddfnl/croot/torchvision_1746818818067/work
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions @ file:///C:/b/abs_ecq8gc0vbm/croot/typing_extensions_1756281142218/work
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug @ file:///C:/b/abs_c7bupijx2_/croot/werkzeug_1737448709012/work
zipp @ file:///C:/b/abs_2fcpcg89my/croot/zipp_1732630746332/work
